[FEAT] : chat activity 구현
